#8638 ExasolIndexConfigurator moved from ext.exasol to ext.exasol.ui
